feat: ensure parent file is always added for decomposed file
